Fix incorrect cache usage by available_properties (big cache use, slow reads)
Spree::Filters::PropertyPresenter includes all ProductProperty for all Products, which can be very many.
This would be serialized and stored in the cache by Rails.cache.fetch.
As only uniq_values are actually relevant and being used, all the ProductProperty's should not be stored in cache.
The number of these properties can be very significant (in the 1000s), taking up a lot of cache and making
the cache read very slow (over 1 second with solid_cache).