Do not apply active class to currently active nav bar menu, because the header (and main nav bar) is fragment cached in production.
Therefore, the same one menu will always have the active class until the cache is invalidated.
This could be better solved with frontend javascript, but just disabling this for now as it's not that important
and doesn't work perfectly anyway (for example the main menu will not be marked active when we are in a submenu).